+++ Packages/webxr-interactions/CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Docs page.
+- WebXR Input Profiles Loader support.
+
+### Changed
+- ControllerInteraction handles hand joints.
 
 ## [0.4.2] - 2020-11-28
 ### Changed