+++ Packages/webxr-interactions/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - OnDisable in ControllerInteraction.
+- Missing material in default WebXRCameraSet.
 
 ## [0.5.1] - 2020-12-26
 ### Fixed